Refactor drag-drop logic to a DragDropContext component
